:white_check_mark: refactor: test Code를 재활용하여 Google Drive 코드 재구성
